pyproject.toml
[project]
name = "langgraph-base"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.10.1",
    "googlemaps>=4.10.0",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.67",
    "langchain-google-genai>=2.1.6",
    "langchain-groq>=0.3.5",
    "langchain-mcp-adapters>=0.1.8",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.0",
    "protobuf==3.20.3",
    "requests>=2.32.4",
    "setuptools>=80.9.0",
    "streamlit>=1.46.1",
    "tavily-python>=0.7.9",
]
